Remove superfluous BSRR code
Learned via the interwebs that, for a GPIO port,

- ODR updates all bits
  (and therefore requires a separate read
  if one wants to update a single bit);
- BRR sets all bits that are 1 to 1;
- BSRR sets all bits that are 0 to 0.

Therefore this commit removes the 'BSRR with all 0s' code,
which is for now superfluous, since
immediately afterwards ODR is used to set all bits.

Later BRR and BSRR might come in handy,
if one wants to set/reset a single bit
without knowing the values of the other bits.
But it looks like svd4zig's generated code
currently does a read anyway?